==> Pandas/1 - Base de Dados.ipynb <==
# Relatório de análise I
## Importando a base de dados

import pandas as pd

alugueis = pd.read_csv('dados/aluguel.csv', sep=";")

alugueis.sample(5)

## Informações gerais sobre a base de dados

tipos_dado_alugueis = pd.DataFrame(alugueis.dtypes, columns = ["Tipos de dados"])
tipos_dado_alugueis.columns.name = "Variáveis"
tipos_dado_alugueis

shape = alugueis.shape
print(f"A base de dados apresenta {shape[0]} registros (imóveis) e {shape[1]} variáveis")
