==> Pandas/2 - Tipos de Imóveis.ipynb <==
# Relatório de Análise II
## Tipos de Imóveis

### Preparação dos dados

import pandas as pd

dados = pd.read_csv('dados/aluguel.csv', sep=";")
dados.sample(5)

### Manipulação dos dados

dados.Tipo

# Retorna uma Series
tipos_imovel = dados.Tipo

# Poderia ser substituido pelo unique
# O Inplace define que o array em questão será modificado
tipos_imovel.drop_duplicates(inplace=True)
tipos_imovel

### Organizando a visualização

# Transformando a series em DataFrame
tipos_imovel = pd.DataFrame(tipos_imovel)

tipos_imovel

# O index está bagunçado
tipos_imovel.index

qtd_indices = tipos_imovel.shape[0]
qtd_indices

tipos_imovel.index = range(qtd_indices)
tipos_imovel.columns.name = "Id"
tipos_imovel
