==> Pandas/3 - Imóveis Residenciais.ipynb <==
# Relatório de Análise III
## Imóveis residenciais
### Preparação do ambiente

import pandas as pd

alugueis = pd.read_csv('dados/aluguel.csv', sep=";")
alugueis.sample(10)

### Preparação dos dados

lista_tipos = list(alugueis.Tipo.drop_duplicates())
lista_tipos

residencial = ['Quitinete',
 'Casa',
 'Apartamento',
 'Casa de Condomínio',
 'Flat',
 'Casa de Vila',
 'Loft']

e_residencial = alugueis.Tipo.isin(residencial)

alugueis_residenciais = alugueis[e_residencial]

alugueis_residenciais.sample(5)

alugueis_residenciais.describe()

list(alugueis_residenciais.Tipo.drop_duplicates())

alugueis_residenciais.shape

alugueis.shape

alugueis.shape[0] - alugueis_residenciais.shape[0]

alugueis_residenciais.index = range(alugueis_residenciais.shape[0])
alugueis_residenciais.head()

### Exportando os dados

alugueis_residenciais.to_csv('dados/alugueis_residenciais.csv', sep = ';')

imp_alugueis_residenciais = pd.read_csv('dados/alugueis_residenciais.csv', sep = ';')

# Foi exportado com o índice do DataFrame
imp_alugueis_residenciais

# Para remover a exportação do índice automático, basta:
alugueis_residenciais.to_csv('dados/alugueis_residenciais.csv', sep = ';', index = False)

imp_alugueis_residenciais = pd.read_csv('dados/alugueis_residenciais.csv', sep = ';')

# Exportado agora sem os índices
imp_alugueis_residenciais
