==> Pandas/4 - Seleções e Frequências.ipynb <==
# Relatório de Análise IV
## Seleções e Frequências
### Preparando o ambiente

import pandas as pd

alugueis_residenciais = pd.read_csv('dados/alugueis_residenciais.csv', sep = ';')
alugueis_residenciais.sample(10)

### Executando Requisitos

#### Selecione somente os imóveis classificados com tipo 'Apartamento'

alugueis_apartamentos = alugueis_residenciais.query('Tipo == "Apartamento"')
alugueis_apartamentos

#### Selecione os imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'

tipos_casa = ['Casa', 'Casa de Condomínio', 'Casa de Vila']
alugueis_casas = alugueis_residenciais[alugueis_residenciais.Tipo.isin(tipos_casa)]
alugueis_casas.sample(10)

#### Selecione os imóveis com área entre 60 e 100 metros quadrados, incluindo os limites

alugueis_60_100 = alugueis_casas[alugueis_casas["Area"].between(60, 100)]
alugueis_60_100

#### Selecione os imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00

alugueis_4_quartos_2_mil = alugueis_casas.query('Quartos >= 4 and Valor < 2000')
alugueis_4_quartos_2_mil

### Exercícios

alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6], 
                        'Aprovado': [True, False, False, True, True, True, False, False]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas', 'Aprovado'])
alunos

#### Apenas alunos aprovados

# Minha resposta
aprovados = alunos[alunos['Aprovado']]
aprovados

# Resposta curso
selecao = alunos['Aprovado'] == True
aprovados = alunos[selecao]
aprovados

#### Apenas alunas aprovados (aprovados e sexo feminino)

# Minha resposta
alunas_aprovadas = alunos[(alunos['Aprovado']) & (alunos['Sexo'] == 'F')]
alunas_aprovadas

# Resposta curso
selecao = (alunos.Aprovado == True) & (alunos.Sexo == 'F')
alunas_aprovadas = alunos[selecao]
alunas_aprovadas

#### Alunos com idade entre 10 e 20 anos ou com idade maior ou igual a 40 anos

# Minha resposta
selecao = (alunos.Idade.between(10, 20)) | (alunos.Idade >= 40)
alunos_idade = alunos[selecao]
alunos_idade

# Resposta curso
selecao = (alunos.Idade > 10) & (alunos.Idade < 20) | (alunos.Idade >= 40)
alunos_idade = alunos[selecao]
alunos_idade

#### Alunos reprovados e apenas as colunas Nome, Sexo e Idade

# Minha resposta
colunas_selecao = ['Nome', 'Sexo', 'Idade']
reprovados = alunos[(alunos.Aprovado == False)][colunas_selecao]
reprovados

# Respostas curso

# Resposta 1
selecao = alunos['Aprovado'] == False
# É possível passar um filtro no loc
reprovados = alunos.loc[selecao, ['Nome', 'Sexo', 'Idade']]
reprovados

# Resposta 2
selecao = alunos['Aprovado'] == False
# A ordem pode ser invertida, mesmo que o filtro não seja executado em uma coluna selecionada
reprovados = alunos[['Nome', 'Sexo', 'Idade']][selecao]
reprovados

#### Os três alunos mais novos

# Minha resposta
mais_novos = alunos.sort_values(by = "Idade")[:3]
mais_novos

# Resposta curso
alunos.sort_values(by = ['Idade'], inplace = True)
alunos.iloc[:3]
