==> Pandas/5 - Tratamento de Dados Faltantes.ipynb <==
# Relatório de Análise V
## Tratamento de dados faltantes
### Preparando o ambiente

import pandas as pd

alugueis_residenciais = pd.read_csv('dados/alugueis_residenciais.csv', sep = ';')
alugueis_residenciais.sample(10)

### Executando o tratamento de dados

# Retorna um DataFrame booleano para cada valor, se é nulo ou não
alugueis_residenciais.isnull()

# Retorna um DataFrame inverso do anterior
alugueis_residenciais.notnull()

# Registros totais vs Registros não-nulos
alugueis_residenciais.info()

ar_valor_nulo = alugueis_residenciais[alugueis_residenciais.Valor.isnull()]
ar_valor_nulo

# No caso da coluna valor, os registros sem esta coluna preenchida não são válidos para nosso conjunto de dados
shape_inicial = alugueis_residenciais.shape[0]
alugueis_residenciais.dropna(subset = ['Valor'], inplace = True)
shape_final = alugueis_residenciais.shape[0]

print(f'Valores removidos: {shape_inicial - shape_final}')

# Não retorna mais dados
ar_valor_nulo = alugueis_residenciais[alugueis_residenciais.Valor.isnull()]
ar_valor_nulo

### Tratamento condicional

alugueis_residenciais[alugueis_residenciais['Condominio'].isnull()]

ar_valor_nulo = alugueis_residenciais[alugueis_residenciais.Condominio.isnull()]
ar_valor_nulo

# Alugueis de apartamento com condomínios nulos também devem ser descartados
com_condominio = ["Apartamento", "Casa de Condomínio"]
selecao = (alugueis_residenciais.Tipo.isin(com_condominio)) & (alugueis_residenciais.Condominio.isnull())
alugueis_residenciais[selecao]

shape_inicial = alugueis_residenciais.shape[0]
#Inverte a seleção
alugueis_residenciais  = alugueis_residenciais[~selecao]
shape_final = alugueis_residenciais.shape[0]

print(f'Valores removidos: {shape_inicial - shape_final}')

alugueis_residenciais

alugueis_residenciais.IPTU.fillna(0, inplace = True)
alugueis_residenciais

alugueis_residenciais.info()

# Condomínios nulos são os que não são de imóveis de condomínio. Tratando, então:
alugueis_residenciais.Condominio.fillna(0, inplace = True)
alugueis_residenciais.info()

#### Exportando os dados

alugueis_residenciais.to_csv('dados/alugueis_residenciais_filtrados.csv', sep = ';', index = False)
