==> Pandas/6 - Novas Variáveis.ipynb <==
# Relatório de Análise VI
## Criando novas variáveis
### Preparando o ambiente

import pandas as pd

alugueis_residenciais = pd.read_csv('dados/alugueis_residenciais_filtrados.csv', sep = ';')
alugueis_residenciais.sample(10)

### Tratando os dados
#### Valor Bruto

alugueis_residenciais['Valor Bruto'] = alugueis_residenciais.Valor + alugueis_residenciais.Condominio + alugueis_residenciais.IPTU

alugueis_residenciais

#### Valor do m²

alugueis_residenciais['Valor m2'] = alugueis_residenciais.Valor / alugueis_residenciais.Area
alugueis_residenciais['Valor Bruto m2'] = alugueis_residenciais['Valor Bruto'] / alugueis_residenciais.Area

alugueis_residenciais['Valor m2'] = alugueis_residenciais['Valor m2'].round(2)
alugueis_residenciais['Valor Bruto m2'] = alugueis_residenciais['Valor Bruto m2'].round(2)
alugueis_residenciais

#### Agrupamento por tipos

alugueis_residenciais.Tipo.drop_duplicates()

casa = ['Casa', 'Casa de Condomínio', 'Casa de Vila']
#apartamento = ['Quitinete', 'Apartamento', 'Flat', 'Loft']

alugueis_residenciais['Tipo Agregado'] = alugueis_residenciais['Tipo'].apply(lambda x: 'Casa' if x in casa else 'Apartamento')
alugueis_residenciais.sample(10)

## Excluindo Variáveis

aux = pd.DataFrame(alugueis_residenciais[['Tipo', 'Valor Bruto', 'Area']])
aux

del aux['Tipo']

aux

_O curso indica a exclusão do dado bruto devido ao IPTU nem sempre ser válido, mas eu não removerei este dado por considerar relevante e ter tratado o IPTU anteriormente_

## Exportando Dados

alugueis_residenciais.to_csv('dados/t_alugueis_residenciais.csv', sep = ';', index = False)
