==> Pandas/8.1 - Exercíclio Box-plot.ipynb <==
# Exercício box-plot

%matplotlib inline
import pandas as pd
import matplotlib.pyplot as plt
plt.rc('figure', figsize = (14, 6))

alugueis = pd.read_csv('dados/aluguel_amostra.csv', sep = ';')
alugueis.head(10)

valor_m2 = alugueis['Valor m2']

type(valor_m2)

Q1 = valor_m2.quantile(.25)
Q3 = valor_m2.quantile(.75)
IIQ = Q3 - Q1
LI = Q1 - 1.5 * IIQ
LS = Q3 + 1.5 * IIQ

print(Q1)
print(Q3)
print(IIQ)
print(LI)
print(LS)
